Lower log level for TTL extraction to warn.
These aren't fatal errors and servers can give back all sorts of wacky
strings. The default values should handle these cases fine, so some
people may not want the noisy error in their logs. Lowering the
severity makes that easier and the exception level is equivalent to
ERROR which is not something that should be silenced.

Fixes #76.