Rewrite functionality removed from cachetools 3
For some reason, cachetools version 3 (released on 2018-11-04) removed
the ability to pass in a `missing` function; apparently this made
the implementation cleaner (see
https://github.com/tkem/cachetools/issues/58). Unfortunately we
use this functionality here.

Rather than require version 2 of cachetools, I thought it would be
better to allow reppy to be used with the new version. Unfortunately,
that means a reimplementation of pretty much exactly what was removed
from cachetools.

I found this because we have a project which uses reppy and its tests
suddenly broke, because pip automatically installed the newer version of
cachetools.